requirements.txt
maubot
matrix-nio[e2e]
sqlitedict